requirements.txt
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
exceptiongroup==1.3.0
fastapi==0.115.12
h11==0.16.0
idna==3.10
importlib_metadata==8.7.0
packaging==25.0
pefile==2023.2.7
pillow==11.2.1
pydantic==2.11.4
pydantic_core==2.33.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
pystray==0.19.5
python-multipart==0.0.20
pywin32-ctypes==0.2.3
qrcode==8.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
zipp==3.21.0
